Add option to specify time to wait before force-killing a worker

+++ cluster-master.js
@@ -296,10 +296,12 @@ function forkListener () {
     worker.on("disconnect", function () {
       debug("Worker %j disconnect", id)
       // give it 1 second to shut down gracefully, or kill
-      disconnectTimer = setTimeout(function () {
-        debug("Worker %j, forcefully killing", id)
-        worker.process.kill("SIGKILL")
-      }, 5000)
+      if(config.killWaitTime !== 0) {
+        disconnectTimer = setTimeout(function () {
+          debug("Worker %j, forcefully killing", id)
+          worker.process.kill("SIGKILL")
+        }, config.killWaitTime || 5000)
+      }
     })
   })
 }
